requirements.txt
Flask==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
Werkzeug==3.0.1
transformers==4.39.2
pillow==10.2.0
torch==2.2.2
pytest==8.1.1
ImageHash==4.3.1
pdf2image==1.17.0
PyMuPDF==1.24.1